+++ gcc/testsuite/gfortran.dg/g77/19981216-0.f
@@ -24,7 +24,7 @@ c { dg-do compile }
 * -------------------------------------------
         PROGRAM WAP
 
-        integer*2  ios
+        integer(kind=8)  ios
         character*80  name
 
         name = 'blah'
@@ -34,7 +34,7 @@ c { dg-do compile }
       END
 * -------------------------------------------
 * 
-* The problem seems to be caused by the "integer*2 ios" declaration.
+* The problem seems to be caused by the "integer(kind=2) ios" declaration.
 * So far I solved it by simply using a plain integer instead.
 * 
 * I'm running gcc on a Linux system compiled/installed
